+++ src/helpers/ethRpc.js
@@ -41,7 +41,6 @@ export const rpcGasPrice = () => JSON.stringify({
 // free nodes without API keys from https://ethereumnodes.com/
 export const ETH_NODES = [
   'https://api.mycryptoapi.com/eth', // MyCrypto
-  'https://web3.1inch.exchange/', // 1inch
   'https://cloudflare-eth.com/', // Cloudflare
   'https://mainnet-nethermind.blockscout.com/', // Blockscout
   'https://nodes.mewapi.io/rpc/eth', // MyEtherWallet