+++ src/xdai/token.js
@@ -1,37 +1,34 @@
 // https://blockscout.com/poa/xdai/bridged-tokens
-export const TOKENMAP = {
-  "0xDDAfbb505ad214D7b80b1f830fcCc89B60fb7A83": {
-    "name": "USD Coin",
-    "logo": "usdc.png",
-    "erc20": true,
-    "symbol": "USDC",
-    "decimals": 6
+// {"chainId": 1, "address": "0x00000000441378008ea67f4284a57932b1c000a5", "name": "TrueGBP", "symbol": "TGBP", "decimals": 18, "tags": []},
+export const TOKENMAP = [
+  {
+    address: '0xDDAfbb505ad214D7b80b1f830fcCc89B60fb7A83',
+    name: 'USD Coin',
+    symbol: 'USDC',
+    decimals: 6,
   },
-  "0x6A023CCd1ff6F2045C3309768eAd9E68F978f6e1": {
-    "name": "Wrapped Ether",
-    "logo": "eth.png",
-    "erc20": true,
-    "symbol": "WETH",
-    "decimals": 18
+  {
+    address: '0x6A023CCd1ff6F2045C3309768eAd9E68F978f6e1',
+    name: 'Wrapped Ether',
+    symbol: 'WETH',
+    decimals: 18,
   },
-  "0x4ECaBa5870353805a9F068101A40E0f32ed605C6": {
-    "name": "Tether USD",
-    "logo": "tether_usd.png",
-    "erc20": true,
-    "symbol": "USDT",
-    "decimals": 6
+  {
+    address: '0x4ECaBa5870353805a9F068101A40E0f32ed605C6',
+    name: 'Tether USD',
+    symbol: 'USDT',
+    decimals: 6,
   },
-  "0x44fA8E6f47987339850636F88629646662444217": {
-    "name": "Dai Stablecoin",
-    "logo": "dai.svg",
-    "erc20": true,
-    "symbol": "DAI",
-    "decimals": 18
+  {
+    address: '0x44fA8E6f47987339850636F88629646662444217',
+    name: 'Dai Stablecoin',
+    symbol: 'DAI',
+    decimals: 18,
   },
-  "0x4Bd5a011E3Dd015106A6f1326a5ba3378610Cb3b": {
-    "name": "Circles",
-    "erc20": true,
-    "symbol": "CRC",
-    "decimals": 18,
-  }
-};
+  {
+    address: '0x4Bd5a011E3Dd015106A6f1326a5ba3378610Cb3b',
+    name: 'Circles',
+    symbol: 'CRC',
+    decimals: 18,
+  },
+];